fix(types): annotate relationships as potentially undefined (#158)
* fix(types): relationships can be undefined

When a model is created any property that does not have an initial value
uses the factory method to create one. Since the factories are typed to
only allow certain primitive values it is not possible for those values
to ever be undefined. For relational properties created through oneOf
and manyOf this is not the case however, if they are not provided an
initial value they are undefined. Currently the types of relational
properties do not include undefined as a possible type.

* fix(relations): fix potentially undefined properties access

Co-authored-by: Artem Zakharchenko <kettanaito@gmail.com>