fix(executeQuery): support circular references when updating entities (#163)
* test: bi-directional relationship support

Add test cases for bi-directional relationships: relationships where
both models are related to each other. An example of this type of
relationship might be a User that has many Post models in `posts`
while the Post model has one User in `author`.

* fix: support circular references when updating entities

The logs for executeQuery JSON.stringify the query. Since the query
includes the data passed that means that any entities passed as data are
also stringified. When an entity has a populated bi-directional
relationship it contains circular references to itself, and so causes a
"cyclic value" exception to be thrown when stringifying.

Use a `safeStringify` util that replaces circular references with string
representations of the entity instead of JSON.stringify.

* chore: polish test, annotate return of "safeStringify"

Co-authored-by: Artem Zakharchenko <kettanaito@gmail.com>